+++ VocaMemo/VocaMemo/Word/WordSheetView.swift
@@ -1,0 +1,15 @@
+//
+//  WordSheetView.swift
+//  VocaMemo
+//
+//  Created by 이영준 on 2022/05/01.
+//
+
+import SwiftUI
+
+struct WordSheetView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
