--- VocaMemo/VocaMemo/WordView.swift
@@ -1,26 +1,0 @@
-//
-//  Word.swift
-//  VocaMemo
-//
-//  Created by 이영준 on 2022/04/30.
-//
-
-import SwiftUI
-
-struct WordView: View {
-    var body: some View {
-        VStack{
-            Text("Word List")
-            
-            
-            
-            
-        }
-    }
-}
-
-struct WordView_Previews: PreviewProvider {
-    static var previews: some View {
-        WordView()
-    }
-}
